[IMP] partner_invoicing_mode_at_shipping: Refactor tests and change to 'shipping' instead of 'picking'
