+++ partner_invoicing_mode_at_shipping/README.rst
@@ -7,7 +7,7 @@ Partner Invoicing Mode At Shipping
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:c3631731c8b2841936e312dd563cdb6e41134c558ffc530d9675123065346d96
+   !! source digest: sha256:47e9c7cd8fac525769bb0981e527ee62f1fa76548509eff6a0c41cecd5222ee8
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -33,7 +33,7 @@ It is based on `partner_invoicing_mode`.
 When this mode is selected the customer will be invoiced automatically on
 delivery of the goods.
 
-Another option is the 'One Invoice Per Picking'. That one is not compatible
+Another option is the 'One Invoice Per Shipping'. That one is not compatible
 with the 'At shipping' invoicing mode. In that case, the invoicing validation
 will occur at a different moment (monthly, ...).
