+++ partner_invoicing_mode_at_shipping/models/stock_picking.py
@@ -19,7 +19,7 @@ class StockPicking(models.Model):
         self.ensure_one()
         return self.picking_type_code == "outgoing" and (
             self.sale_id.partner_invoice_id.invoicing_mode == "at_shipping"
-            or self.sale_id.partner_invoice_id.one_invoice_per_picking
+            or self.sale_id.partner_invoice_id.one_invoice_per_shipping
         )
 
     def _invoicing_at_shipping(self):
@@ -37,7 +37,7 @@ class StockPicking(models.Model):
             invoices |= sales_many_invoice_per_order._create_invoices(grouped=False)
         # The invoices per picking will use the invoicing_mode
         for invoice in invoices.filtered(
-            lambda invoice: not invoice.partner_id.one_invoice_per_picking
+            lambda invoice: not invoice.partner_id.one_invoice_per_shipping
         ):
             invoice.with_delay()._validate_invoice()
         return invoices or _("Nothing to invoice.")
