+++ partner_invoicing_mode_at_shipping/static/description/index.html
@@ -367,14 +367,14 @@ ul.auto-toc {
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-!! source digest: sha256:c3631731c8b2841936e312dd563cdb6e41134c558ffc530d9675123065346d96
+!! source digest: sha256:47e9c7cd8fac525769bb0981e527ee62f1fa76548509eff6a0c41cecd5222ee8
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
 <p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/OCA/account-invoicing/tree/16.0/partner_invoicing_mode_at_shipping"><img alt="OCA/account-invoicing" src="https://img.shields.io/badge/github-OCA%2Faccount--invoicing-lightgray.png?logo=github" /></a> <a class="reference external image-reference" href="https://translation.odoo-community.org/projects/account-invoicing-16-0/account-invoicing-16-0-partner_invoicing_mode_at_shipping"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external image-reference" href="https://runboat.odoo-community.org/builds?repo=OCA/account-invoicing&amp;target_branch=16.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
 <p>This module allows to select a <cite>At shipping</cite> invoicing mode for a customer.
 It is based on <cite>partner_invoicing_mode</cite>.
 When this mode is selected the customer will be invoiced automatically on
 delivery of the goods.</p>
-<p>Another option is the ‘One Invoice Per Picking’. That one is not compatible
+<p>Another option is the ‘One Invoice Per Shipping’. That one is not compatible
 with the ‘At shipping’ invoicing mode. In that case, the invoicing validation
 will occur at a different moment (monthly, …).</p>
 <p><strong>Table of contents</strong></p>
