+++ partner_invoicing_mode_at_shipping/tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_invoice_mode_at_shipping
+from . import test_invoice_mode_at_shipping, test_invoice_mode_group_delivery
