+++ partner_invoicing_mode_at_shipping/tests/common.py
@@ -1,0 +1,56 @@
+# Copyright 2023 ACSONE SA/NV
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+
+class InvoiceModeAtShippingCommon:
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.env = cls.env(context=dict(cls.env.context, tracking_disable=True))
+        cls.partner = cls.env.ref("base.res_partner_1")
+        cls.product = cls.env.ref("product.product_delivery_01")
+        cls.so1 = cls.env["sale.order"].create(
+            {
+                "partner_id": cls.partner.id,
+                "partner_invoice_id": cls.partner.id,
+                "partner_shipping_id": cls.partner.id,
+                "order_line": [
+                    (
+                        0,
+                        0,
+                        {
+                            "name": "Line one",
+                            "product_id": cls.product.id,
+                            "product_uom_qty": 4,
+                            "product_uom": cls.product.uom_id.id,
+                            "price_unit": 123,
+                        },
+                    )
+                ],
+                "pricelist_id": cls.env.ref("product.list0").id,
+            }
+        )
+
+    @classmethod
+    def _create_order(cls):
+        cls.so1 = cls.env["sale.order"].create(
+            {
+                "partner_id": cls.partner.id,
+                "partner_invoice_id": cls.partner.id,
+                "partner_shipping_id": cls.partner.id,
+                "order_line": [
+                    (
+                        0,
+                        0,
+                        {
+                            "name": "Line one",
+                            "product_id": cls.product.id,
+                            "product_uom_qty": 4,
+                            "product_uom": cls.product.uom_id.id,
+                            "price_unit": 123,
+                        },
+                    )
+                ],
+                "pricelist_id": cls.env.ref("product.list0").id,
+            }
+        )
